+++ addons/godot_editor_discord_presence/Discord RPC/ipc/module/rich presence/RichPresence.gd
@@ -1,5 +1,7 @@
 class_name RichPresence
 
+# warning-ignore-all:narrowing_conversion
+
 var state: String
 var details: String
 var start_timestamp: int