+++ addons/godot_editor_discord_presence/Discord RPC/ipc/pipe/UnixPipe.gd
@@ -7,19 +7,24 @@ func open(path: String) -> int:
 	return _peer.open(path)
 
 func read() -> Array:
+	if not is_open():
+		return [-1, PoolByteArray()]
+	
 	var op_code: int = _peer.get_32()
 	var length: int = _peer.get_32()
 	var buffer: PoolByteArray = _peer.get_data(length)[1]
+	
 	return [op_code, buffer]
 
 func write(bytes: PoolByteArray) -> void:
-	_peer.put_data(bytes)
+	if is_open():
+		_peer.put_data(bytes)
 
 func is_open() -> bool:
 	return _peer and _peer.is_open()
 
 func has_reading() -> bool:
-	return _peer.get_available_bytes() > 0
+	return _peer.get_available_bytes() > 0 if _peer else false
 
 func close() -> void:
 	if _peer and _peer.is_open():