+++ addons/godot_editor_discord_presence/Discord RPC/util/IPCUtil.gd
@@ -3,14 +3,14 @@ class_name IPCUtil
 class HandshakePayload extends IPCPayload:
 	var version: int
 	var client_id: int
-
+	
 	# warning-ignore:shadowed_variable
 	# warning-ignore:shadowed_variable
 	func _init(version: int, client_id: int) -> void:
 		self.op_code = OpCodes.HANDSHAKE
 		self.version = version
 		self.client_id = client_id
-
+	
 	func to_dict() -> Dictionary:
 		return {
 			"v": self.version,
@@ -20,7 +20,7 @@ class HandshakePayload extends IPCPayload:
 class AuthorizePayload extends IPCPayload:
 	func _init(client_id: int, scopes: PoolStringArray) -> void:
 		self.op_code = OpCodes.FRAME
-		self.command = DiscordRPCEnum.Commands.AUTHORIZE
+		self.command = DiscordRPCUtil.Commands.AUTHORIZE
 		self.arguments = {
 			"client_id": str(client_id),
 			"scopes": scopes
@@ -29,19 +29,19 @@ class AuthorizePayload extends IPCPayload:
 class AuthenticatePayload extends IPCPayload:
 	func _init(access_token: String) -> void:
 		self.op_code = OpCodes.FRAME
-		self.command = DiscordRPCEnum.Commands.AUTHENTICATE
+		self.command = DiscordRPCUtil.Commands.AUTHENTICATE
 		self.arguments = {"access_token": access_token}
 
 class SubscribePayload extends IPCPayload:
 	func _init(subscribe_event: String, arg: Dictionary) -> void:
 		self.op_code = OpCodes.FRAME
-		self.command = DiscordRPCEnum.Commands.SUBSCRIBE
+		self.command = DiscordRPCUtil.Commands.SUBSCRIBE
 		self.arguments = arg
 		self.event = subscribe_event.to_upper()
 
 class UnsubscribePayload extends IPCPayload:
 	func _init(usubscribe_event: String, arg: Dictionary) -> void:
 		self.op_code = OpCodes.FRAME
-		self.command = DiscordRPCEnum.Commands.UNSUBSCRIBE
+		self.command = DiscordRPCUtil.Commands.UNSUBSCRIBE
 		self.arguments = arg
 		self.event = usubscribe_event.to_upper()