+++ addons/godot_editor_discord_presence/Discord RPC/util/URLUtil.gd
@@ -1,7 +1,5 @@
 class_name URLUtil
 
 static func dict_to_url_encoded(data: Dictionary) -> String:
-	var parameters: PoolStringArray
-	for key in data.keys():
-		parameters.append("%s=%s" % [str(key), str(data[key]).percent_encode()]) 
-	return parameters.join("&")
+	var client: HTTPClient = HTTPClient.new()
+	return client.query_string_from_dict(data)