+++ CHANGELOG.md
@@ -4,6 +4,15 @@ Changelog
 This is a high-level changelog for each released versions of the plugin.
 For a more detailed list of past changes, see the commit history.
 
+1.2.0
+------
+- Updateed DiscordRPC to v1.1.0 and godot-unix-socket to v1.1.0, this fixes a bug where Godot crashes if Discord is closed when connected via DiscordRPC
+
+1.1.0
+------
+- Added auto-reconnect to Discord client
+- Added Linux / OSx support
+
 1.0.0
 ------
 - Initial version